services/avatar-worker/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
opencv-python>=4.8.0
trimesh>=4.0.0
pygltflib>=1.16.0
smplx>=0.1.28
redis>=5.0.0
minio>=7.2.0
requests>=2.31.0
Pillow>=10.0.0
scikit-image>=0.21.0
scipy>=1.11.0
python-dotenv>=1.0.0
braceexpand>=0.1.7
roma>=1.5.0
omegaconf>=2.3.0
yacs>=0.1.8
einops>=0.7.0
timm>=0.9.12
pytorch-lightning>=2.1.0

